HW-4 MovieHistoryContract, MovieHistoryPresenter 생성 및 액티비티에 적용
