+++ class15/dktsudgg/AndroidArchitectureStudy/app/src/main/java/kr/dktsudgg/androidarchitecturestudy/MovieSearchHistoryActivity.kt
@@ -5,23 +5,22 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.activity_movie_search_history.*
-import kr.dktsudgg.androidarchitecturestudy.data.model.MovieItem
 import kr.dktsudgg.androidarchitecturestudy.data.repository.NaverMovieRepositoryImpl
 import kr.dktsudgg.androidarchitecturestudy.databinding.ActivityMovieSearchHistoryBinding
 import kr.dktsudgg.androidarchitecturestudy.view.adapter.MovieSearchHistoryListAdapter
-import kr.dktsudgg.androidarchitecturestudy.view.ui.MovieSearchContract
-import kr.dktsudgg.androidarchitecturestudy.view.ui.MovieSearchPresenter
+import kr.dktsudgg.androidarchitecturestudy.view.ui.MovieHistoryContract
+import kr.dktsudgg.androidarchitecturestudy.view.ui.MovieHistoryPresenter
 
-class MovieSearchHistoryActivity : BaseActivity<MovieSearchPresenter>(), MovieSearchContract.View {
+class MovieSearchHistoryActivity : BaseActivity<MovieHistoryPresenter>(), MovieHistoryContract.View {
 
     private lateinit var binding: ActivityMovieSearchHistoryBinding
-    private lateinit var movieSearchPresenter: MovieSearchContract.Presenter
+    private lateinit var movieHistoryPresenter: MovieHistoryContract.Presenter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, R.layout.activity_movie_search_history)
 
-        movieSearchPresenter = MovieSearchPresenter(NaverMovieRepositoryImpl(), this)
+        movieHistoryPresenter = MovieHistoryPresenter(NaverMovieRepositoryImpl(), this)
 
         /**
          * 영화 검색이력 리스트 보여주는 RecyclerView에 어댑터 연결 및 목록 구분선 추가
@@ -37,16 +36,11 @@ class MovieSearchHistoryActivity : BaseActivity<MovieSearchPresenter>(), MovieSe
         /**
          * 로컬 저장소로부터 영화 검색 이력 조회해옴
          */
-        movieSearchPresenter.showMovieSearchHistory()
+        movieHistoryPresenter.showMovieSearchHistory()
     }
 
     override fun updateMovieSearchHistoryList(data: List<String>) {
         (movieSearchHistoryList.adapter as MovieSearchHistoryListAdapter).refreshData(data)
     }
-
-    override fun updateSearchedMovieList(data: List<MovieItem>) {
-        TODO("Not yet implemented")
-    }
-
 
 }
