+++ class15/dktsudgg/AndroidArchitectureStudy/app/src/main/java/kr/dktsudgg/androidarchitecturestudy/view/ui/MovieHistoryContract.kt
@@ -1,0 +1,19 @@
+package kr.dktsudgg.androidarchitecturestudy.view.ui
+
+interface MovieHistoryContract {
+
+    interface View {
+        /**
+         * 영화 검색이력 조회결과 갱신 메소드
+         */
+        fun updateMovieSearchHistoryList(data: List<String>)
+    }
+
+    interface Presenter {
+        /**
+         * 영화 검색이력 조회 메소드
+         */
+        fun showMovieSearchHistory()
+    }
+
+}
