+++ class15/dktsudgg/AndroidArchitectureStudy/app/src/main/java/kr/dktsudgg/androidarchitecturestudy/view/ui/MovieHistoryPresenter.kt
@@ -1,0 +1,16 @@
+package kr.dktsudgg.androidarchitecturestudy.view.ui
+
+import kr.dktsudgg.androidarchitecturestudy.data.repository.NaverMovieRepository
+
+class MovieHistoryPresenter(
+    private val naverMovieRepository: NaverMovieRepository,
+    private val view: MovieHistoryContract.View
+) : BasePresenter(), MovieHistoryContract.Presenter {
+
+    override fun showMovieSearchHistory() {
+        view.updateMovieSearchHistoryList(
+            naverMovieRepository.getMovieSearchHistory()
+        )
+    }
+
+}
