+++ class15/dktsudgg/AndroidArchitectureStudy/app/src/main/java/kr/dktsudgg/androidarchitecturestudy/view/ui/MovieSearchContract.kt
@@ -9,11 +9,6 @@ interface MovieSearchContract {
          * 영화 검색 결과 갱신 메소드
          */
         fun updateSearchedMovieList(data: List<MovieItem>)
-
-        /**
-         * 영화 검색이력 조회결과 갱신 메소드
-         */
-        fun updateMovieSearchHistoryList(data: List<String>)
     }
 
     interface Presenter : BaseContract.Presenter {
@@ -21,11 +16,6 @@ interface MovieSearchContract {
          * 영화 검색 메소드
          */
         fun searchMovies(query: String)
-
-        /**
-         * 영화 검색이력 조회 메소드
-         */
-        fun showMovieSearchHistory()
     }
 
 }
