+++ class15/dktsudgg/AndroidArchitectureStudy/app/src/main/java/kr/dktsudgg/androidarchitecturestudy/view/ui/MovieSearchPresenter.kt
@@ -28,9 +28,4 @@ class MovieSearchPresenter(
         })
     }
 
-    override fun showMovieSearchHistory() {
-        view.updateMovieSearchHistoryList(
-            naverMovieRepository.getMovieSearchHistory()
-        )
-    }
 }
